fix(modal): support close animation
Tests now have to wait for animation before checking for open modals.
Tests should actually check for the 'in' class.

Setting up both the transition end handler and a timeout ensures that
the modal is always closed even if there was an issue with the
transition. This was the implementation used by Twitter Bootstrap modal
JS code.

Note that unbinding the transition end event within the event
handler isn't necessary, and, worse is currently buggy in jqLite (see
https://github.com/angular/angular.js/pull/5109 ).

Note that the scope is already destroyed when the dom is removed so the
$destroy call isn't needed.

Closes #1341