docs(CHANGELOG): Added dropdown breaking change
After I upgraded from 0.10.0 to 0.11.0 I needed to add the dropdown directive to my button dropdowns. This is not documented in the breakting changes.

The bootstrap documentation only mentions the dropdown class only in the [dropdowns section](http://getbootstrap.com/components/#dropdowns), but not in the [btn-dropdowns section](http://getbootstrap.com/components/#btn-dropdowns).