fix(tooltip): remove extra digest causing incompatibility
In Angular 1.3.1, when using tooltips with ngAnimate, this extra $digest
call seems to break ngClass. This is an extra call to ngAnimate and it
shouldn’t be called since the tooltip hasn’t been added to the DOM.

I was unable to create a test case to test compatibility with ngAnimate.

Fixes #2951
Fixes #2959