fix(typeahead): reset 'parse' validation key
When typeahead-editable="false" and we select a perfect
match from the results list, the $parsers aren't called,
which results in the 'parse' error key not being reset.

As with the 'editable' key, we should reset this once
an item is selected because we know we have a valid
model matching the view value.

Closes #3166